+++ 02_소스코드/이준민/Front/Service/2024.12.03 11.09/Payment/PaymentService.js
@@ -1,0 +1,29 @@
+// services - DeptService.js
+import axios from "axios";
+// 벡엔드 주소 : springboot 주소(컨트롤러주소)
+const baseURL = "http://localhost:8000/api/advanced"
+
+const getAll = (searchKeyword, pageIndex, recordCountPerPage) => {
+    return axios.get(baseURL+`/payment?searchKeyword=${searchKeyword}&pageIndex=${pageIndex}&recordCountPerPage=${recordCountPerPage}`);
+}
+const get = (paid) => {
+    return axios.get(baseURL+`/payment/${paid}`);
+}
+const insert = (data) => {
+    return axios.post(baseURL+"/payment/insert", data);
+}
+const update = (paid, data) => {
+    return axios.put(baseURL+`/payment/update/${paid}`, data);
+}
+const remove = (paid) => {
+    return axios.delete(baseURL+`/payment/delete/${paid}`);
+}
+const ExampleService = {
+    getAll,
+    get,
+    insert,
+    update,
+    remove 
+}
+
+export default ExampleService;
